Refactor primitives modules to a subdirectory
